+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Features:
 Bugfixes:
 
 * Skip empty values when restoring tables (#348)
+* Only append to table when first value is a float
 * GUI send/receive config for objects in subpatches
 * DPF: also send BPM value to `__hv_dpf_bpm` when transport is not playing